hostfootprint/requirements.txt
Django>=1.8,<2.0
psycopg2
ipaddress
python-nmap
ipython
elasticsearch
pytz
django_countries
geopy
